Refactor package hierachy
