+++ src/main/java/com/study/loge/betchat/utils/enums/ResultState.java
@@ -1,0 +1,7 @@
+package com.study.loge.betchat.utils.enums;
+
+// 클라이언트에 메시지 상태를 알려주는 클래스입니다.
+public enum ResultState {
+    OK,
+    ERROR;
+}
