+++ src/main/java/com/study/loge/betchat/utils/exception/ChatMessageTextException.java
@@ -1,4 +1,4 @@
-package com.study.loge.betchat.exception;
+package com.study.loge.betchat.utils.exception;
 
 // 채팅 메시지 텍스트 예외 클래스입니다.
 public class ChatMessageTextException extends Exception{
