+++ src/main/java/com/study/loge/betchat/utils/exception/LoginException.java
@@ -1,4 +1,4 @@
-package com.study.loge.betchat.exception;
+package com.study.loge.betchat.utils.exception;
 
 // 유저 로그인 단계에서 발생하는 예외입니다.
 public class LoginException extends Exception {
