+++ src/main/java/com/study/loge/betchat/utils/exception/RoomJoinException.java
@@ -1,4 +1,4 @@
-package com.study.loge.betchat.exception;
+package com.study.loge.betchat.utils.exception;
 
 // 채팅방 참여 시 발생하는 에외입니다.
 public class RoomJoinException extends Exception{
