+++ src/main/java/com/study/loge/betchat/utils/validation/messageheader/RoomCreationValidationChecker.java
@@ -1,7 +1,7 @@
 package com.study.loge.betchat.utils.validation.messageheader;
 
 import com.study.loge.betchat.entity.User;
-import com.study.loge.betchat.exception.RoomCreateException;
+import com.study.loge.betchat.utils.exception.RoomCreateException;
 import com.study.loge.betchat.model.MessageHeader;
 import com.study.loge.betchat.model.request.RoomCreateRequest;
 import com.study.loge.betchat.repository.UserRepository;
